[refactor]: project structure change
