+++ src/main/java/media/wepg/prototype/controller/response/common/ApiBody.java
@@ -1,0 +1,17 @@
+package media.wepg.prototype.controller.response.common;
+
+import lombok.Getter;
+import org.springframework.web.bind.annotation.ResponseBody;
+
+@Getter
+@ResponseBody
+public class ApiBody<T> {
+
+    private final T data;
+    private final T message;
+
+    public ApiBody(T data, T message) {
+        this.data = data;
+        this.message = message;
+    }
+}
