+++ src/main/java/media/wepg/prototype/controller/response/common/ResponseMessage.java
@@ -1,0 +1,6 @@
+package media.wepg.prototype.controller.response.common;
+
+public class ResponseMessage {
+    public static final String SUCCESS_MESSAGE = "ok";
+    public static final String FAILED_MESSAGE = "failed";
+}
