--- src/main/java/media/wepg/prototype/es/controller/response/common/ResponseMessage.java
@@ -1,6 +1,0 @@
-package media.wepg.prototype.es.controller.response.common;
-
-public class ResponseMessage {
-    public static final String SUCCESS_MESSAGE = "ok";
-    public static final String FAILED_MESSAGE = "failed";
-}
