+++ src/main/java/media/wepg/prototype/model/dto/response/ProgramResponseDto.java
@@ -1,0 +1,41 @@
+package media.wepg.prototype.model.dto.response;
+
+import lombok.Data;
+import media.wepg.prototype.model.es.Program;
+
+import java.time.LocalDateTime;
+
+@Data
+public class ProgramResponseDto {
+
+    private Long serviceId;
+    private String eventId;
+    private LocalDateTime eventStartDate;
+    private LocalDateTime eventEndDate;
+    private String titleName;
+    private String synopsisName;
+    private String directorName;
+    private String actorName;
+    private String songName;
+    private String singerName;
+    private String providerName;
+    private LocalDateTime insertDate;
+    private LocalDateTime updateDate;
+
+    public ProgramResponseDto(Program program) {
+        this.serviceId = program.getServiceId();
+        this.eventId = program.getEventId();
+        this.eventStartDate = program.getEventStartDate();
+        this.eventEndDate = program.getEventEndDate();
+        this.titleName = program.getTitleName();
+        this.synopsisName = program.getSynopsisName();
+        this.directorName = program.getDirectorName();
+        this.actorName = program.getActorName();
+        this.songName = program.getSongName();
+        this.singerName = program.getSingerName();
+        this.providerName = program.getProviderName();
+        this.insertDate = program.getInsertDate();
+        this.updateDate = program.getUpdateDate();
+    }
+
+}
