--- src/main/java/media/wepg/prototype/orig/repository/ChannelRepository.java
@@ -1,9 +1,0 @@
-package media.wepg.prototype.orig.repository;
-
-import media.wepg.prototype.orig.model.ChannelOrigin;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChannelRepository extends JpaRepository<ChannelOrigin, Long> {
-}
