+++ src/test/java/media/wepg/prototype/orig/repository/ChannelOriginRepositoryTest.java
@@ -1,6 +1,7 @@
 package media.wepg.prototype.orig.repository;
 
-import media.wepg.prototype.orig.model.ChannelOrigin;
+import media.wepg.prototype.model.ChannelOrigin;
+import media.wepg.prototype.repository.ChannelRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
